ReceiptsScanner/requirements.txt
# Core Framework
streamlit==1.28.0
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Async Task Processing
celery==5.3.4
redis==5.0.1
flower==2.0.1

# OCR & Image Processing
pytesseract==0.3.10
Pillow==10.1.0
opencv-python==4.8.1.78
pdf2image==1.16.3

# Machine Learning
sentence-transformers==2.2.2
scikit-learn==1.3.2
torch==2.1.0
transformers==4.35.2
mlflow==2.8.1

# Data Processing
pandas==2.1.3
numpy==1.26.2
python-multipart==0.0.6

# Storage
boto3==1.29.7
minio==7.2.0

# Monitoring & Logging
prometheus-client==0.19.0
structlog==23.2.0
sentry-sdk==1.38.0

# Visualization
plotly==5.18.0
matplotlib==3.8.2

# Utilities
python-dotenv==1.0.0
requests==2.31.0
aiofiles==23.2.1
httpx==0.25.2

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
pytest-mock==3.12.0
httpx==0.25.2

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.5.0